% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add}
\alias{add}
\title{The function that adds two numbers}
\usage{
add(A, B)
}
\arguments{
\item{A}{: first  number to add}

\item{B}{: second number to add}
}
\description{
The function that adds two numbers
}
\examples{
add()
}
\keyword{add}
