% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exit_with_error}
\alias{exit_with_error}
\title{The function that exits with an error code and prints an error}
\usage{
exit_with_error(Str)
}
\arguments{
\item{Str}{: String to print upon exit}
}
\description{
The function that exits with an error code and prints an error
}
\examples{
exit_with_error()
}
\keyword{error}
