% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{The main function that calls add()}
\usage{
main(A = 0, B = 0)
}
\arguments{
\item{A}{: first  number to add}

\item{B}{: second number to add}
}
\description{
The main function that calls add()
}
\examples{
main()
}
\keyword{add}
